+++ tests/test_cli.py
@@ -3,6 +3,20 @@ from click.testing import CliRunner
 from influxio.cli import cli
 
 
+def test_info():
+    """
+    CLI test: Invoke `influxio info`
+    """
+    runner = CliRunner()
+
+    result = runner.invoke(
+        cli,
+        args="info",
+        catch_exceptions=False,
+    )
+    assert result.exit_code == 0
+
+
 def test_testdata(caplog):
     """
     CLI test: Import test data to InfluxDB.